Blocknet PoS staking test header
